requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.17.0
fonttools==4.56.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
kiwisolver==1.4.8
langchain-core==0.3.41
langchain-text-splitters==0.3.6
langsmith==0.3.11
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.4.2
numpy==2.2.2
openai==1.63.2
openpyxl==3.1.5
orjson==3.10.15
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.3.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.42.0
svglib==1.5.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tinycss2==1.4.0
together==1.4.1
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
